Reload tailed files on non-linux platforms
Python has an issue on OS X were the underlying C implementation of
`file.read()` caches the EOF, therefore causing `readlines()` to only
work once. This happens to also fail miserably when you are seeking to
the end before calling readlines.

This fix solves the issue by constantly re-reading the files changed.

Note that this also causes debug mode to be very noisy on OS X. We all
have to make sacrifices...