Use GlobSafeConfigParser to parse config files
In order to support all of the kinds of globs, pass GlobSafeConfigParser
into the Configuration object so that it parses section headers
correctly.

Update dependency on conf_d

Fix line spacing + trigger travis

Remove chdir in test

This should fix the unit test to run properly when run from the main
directory.