When shipping logs, use millisecond-precision timestamps.
Logstash 1.3.2 has a problem with microsecond-precision timestamps in the
@timestamp field, which is the default behavior of Python's .isoformat
method. Logstash uses the JodaTime library to parse timestamps, and Joda
doesn’t support nanosecond timestamp resolution. As a result, Logstash
1.3.2 throws an exception on every log item shipped from Beaver.

There's a discussion about this issue in the logstash-users mailing list,
including an example of the Logstash exception:
    https://groups.google.com/forum/#!topic/logstash-users/wIzdv15Iefs

This patch reduces @timestamp to millisecond precision, which should
correct the problem with Beaver 1.3.2.