Add channel support to Redis transport
This changeset adds support for publishing log entries to a Redis
channel, which is also supported by Logstash's Redis input.

Beaver configuration files can now supply a `redis_data_type` key. Valid
values for this key are `list` and `channel`. If left unset, the default
is `list`.

Attempts to resolve #266.