Update http_transport.py
The object edata is a string and when this is sent using 'r = requests.post(url=self._url, data=edata)' no content-type is sent by requests to the receiving web service. If a dict object was set instead requests would detect this and automatically apply a content-type.

This change checks if edata can be decoded as json and if it can adds appropriate header with content-type. If not it still adds a content-type to make the request more robust for the receiving server.

Prior to this change requests was seeing the status_code as a 500, even though it wasn't, because the reply from the web service had no content-type in the header since it didn't know what content-type it had received.